Android 14.0.0 Release 50 (AP2A.240605.024)

# -----BEGIN PGP SIGNATURE-----
#
# iF0EABECAB0WIQRDQNE1cO+UXoOBCWTorT+BmrEOeAUCZmdzpQAKCRDorT+BmrEO
# eJLtAJ43wF6qlq+J5XXdLJ+yWK58/er2DACfacvUKEF8Ioc8D8q6EaNj1laJ59M=
# =oxwW
# -----END PGP SIGNATURE-----
# gpg: Signature made Tue Jun 11 00:44:05 2024 EEST
# gpg:                using DSA key 4340D13570EF945E83810964E8AD3F819AB10E78
# gpg: Good signature from "The Android Open Source Project <initial-contribution@android.com>" [marginal]
# gpg: initial-contribution@android.com: Verified 2480 signatures in the past
#      2 years.  Encrypted 4 messages in the past 2 years.
# gpg: WARNING: This key is not certified with sufficiently trusted signatures!
# gpg:          It is not certain that the signature belongs to the owner.
# Primary key fingerprint: 4340 D135 70EF 945E 8381  0964 E8AD 3F81 9AB1 0E78

# By Andy Hung (28) and others
# Via Automerger Merge Worker (486) and others
* tag 'android-14.0.0_r50': (357 commits)
  MPEG4Writer: Remove timestamp validation
  VTS: check ro.vendor.api_level instead of ro.board.(first_)api_level
  Revert "Introduce AttributionAndPermissionUtils."
  Fix playbackHeadPosition after starting tracks with a startThresholdInFrames
  Add Java library to query state of VirtualCamera build flag.
  Fail if there's problem with tmp buffer during JPEG capture
  Only add the AIDL manifest decl when using unfrozen
  Remove Android.mk files which are not needed any more.
  Extend output format with downscalable resolutions
  AudioStreamOut: Align with AudioStreamIn
  Improve AudioMix registration/unregistration
  android.media.codec-aconfig-cc: Set double_loadable to true
  EffectBufferHalAidl: remove ashmem mapping
  AudioPolicyService: Make device effect init synchronous
  AudioSystem: Move AudioPolicyService fetch to ServiceHandler
  VtsHalMediaC2: start thread pool
  media.c2 client: fix # of queried param count
  Spatializer: replace parameters with AIDL type
  AudioSystem: Move AudioFlinger fetch to ServiceHandler
  audiosystem_tests: Fix false test failures
  ...

 Conflicts:
	media/codec2/sfplugin/CCodecBufferChannel.cpp
	services/camera/libcameraservice/utils/SessionConfigurationUtils.h

+++ include/media/Interpolator.h
@@ -204,7 +204,7 @@ public:
             mInterpolatorType = interpolatorType;
             return NO_ERROR;
         default:
-            ALOGE("invalid interpolatorType: %d", interpolatorType);
+            ALOGE("invalid interpolatorType: %d", static_cast<int>(interpolatorType));
             return BAD_VALUE;
         }
     }
